v.1.01

+++ src/main/java/net/pnjurassic/world/dimension/jurassic/GenLayerJurassic/GenLayerJurassicDeepOcean.java
@@ -39,7 +39,7 @@ public class GenLayerJurassicDeepOcean extends GenLayer
         {
             for (int j1 = 0; j1 < p_151626_3_; ++j1)
             {
-                this.initChunkSeed((long)(j1 + p_151626_1_), (long)(i1 + p_151626_2_));
+                this.initChunkSeed(j1 + p_151626_1_, i1 + p_151626_2_);
                 int k1 = aint[j1 + 1 + (i1 + 1) * k];
 
                 if (k1 == JURASSIC_OCEAN_SHORE_ID)
